Remove get_items_sql from DatabaseAccess
Its unit tests are removed, but the method is retained because it is used
in the unit tests from _get_items_dict
